+++ package.json
@@ -48,7 +48,7 @@
     "eslint": "eslint src/.",
     "typecheck": "npx tsc",
     "test": "jest",
-    "build:types": "tsc --declaration --declarationMap --outDir dist",
+    "build:types": "tsc --declaration --declarationMap --emitDeclarationOnly --outDir dist",
     "postbuild": "npm run build:types",
     "build": "node esbuild.js",
     "prepare": "npm run build",
@@ -76,10 +76,12 @@
     "eslint-plugin-prettier": "^5.2.1",
     "eslint-plugin-promise": "^7.1.0",
     "fs-extra": "^11.2.0",
+    "glob": "^11.0.0",
     "happy-dom": "^15.7.4",
     "identity-obj-proxy": "^3.0.0",
     "jest": "^29.7.0",
     "jest-environment-jsdom": "^29.7.0",
+    "sass": "^1.80.6",
     "stylelint": "^16.10.0",
     "stylelint-config-standard": "^36.0.1",
     "stylelint-config-standard-scss": "^13.1.0",
